requirements.txt
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
selenium~=4.25.0
python-dotenv~=1.0.1